internal/utils/ether.go
package utils

import (
	"fmt"
	"math/big"
	"strings"

	"github.com/ethereum/go-ethereum/params"
)

func ToEther(wei *big.Int) float64 {
	// https://github.com/ethereum/go-ethereum/issues/21221#issuecomment-805852059
	f := new(big.Float)
	f.SetPrec(236)
	f.SetMode(big.ToNearestEven)
	fWei := new(big.Float)
	fWei.SetPrec(236)
	fWei.SetMode(big.ToNearestEven)
	f.Quo(fWei.SetInt(wei), big.NewFloat(params.Ether))
	res, _ := f.Float64()
	return res
}

func ToWei(amount float64) *big.Int {
	eth := big.NewFloat(amount)
	truncInt, _ := eth.Int(nil)
	truncInt = new(big.Int).Mul(truncInt, big.NewInt(params.Ether))
	fracStr := strings.Split(fmt.Sprintf("%.18f", eth), ".")[1]
	fracStr += strings.Repeat("0", 18-len(fracStr))
	fracInt, _ := new(big.Int).SetString(fracStr, 10)
	wei := new(big.Int).Add(truncInt, fracInt)
	return wei
}
